full avances requerimientos generacion master

+++ database/migrations/2025_06_12_142753_create_blocks_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('code', 20)->unique();
             $table->tinyText('name');
             $table->boolean('has_text')->default(false);
-            $table->unsignedInteger('parent_block_id')->nullable()->index();
+            $table->unsignedBigInteger('parent_block_id')->nullable()->index();
             $table->timestamps();
 
             $table->foreign('level_id')
